Updated BE Methods based on review feedback

+++ mvp_onboarding.Server/Classes/CustomerMethods.cs
@@ -17,10 +17,9 @@ namespace mvp_onboarding.Server.Classes
 
         public async Task<CustomerResponseDto> GetCustomers(int pageNumber, int pageSize, string sortColumn, string sortDirection)
         {
-            var query = _context.Customers.AsQueryable();
-
             try
             {
+                var query = _context.Customers.AsQueryable();
                 if (sortDirection.ToLower() == "asc")
                 {
                     query = query.OrderBy(c => EF.Property<Customer>(c, sortColumn));
@@ -29,60 +28,86 @@ namespace mvp_onboarding.Server.Classes
                 {
                     query = query.OrderByDescending(c => EF.Property<Customer>(c, sortColumn));
                 }
+
+                var totalCount = await query.CountAsync();
+                var customers = await query
+                    .Skip((pageNumber - 1) * pageSize)
+                    .Take(pageSize)
+                    .ToListAsync();
+
+                var customerDtos = customers.Select(c => CustomerMapper.EntityToDto(c)).ToList();
+
+                return new CustomerResponseDto
+                {
+                    Items = customerDtos,
+                    TotalCount = totalCount,
+                    PageSize = pageSize,
+                    CurrentPage = pageNumber
+                };
             }
             catch (InvalidOperationException ex)
             {
                 Console.WriteLine(ex.Message);
+                
+                return new CustomerResponseDto
+                {
+                    Items = new List<CustomerDto>(),
+                    TotalCount = 0,
+                    PageSize = pageSize,
+                    CurrentPage = pageNumber
+                };
             }
 
-            var totalCount = await query.CountAsync();
-            var customers = await query
-                .Skip((pageNumber -1) * pageSize)
-                .Take(pageSize)
-                .ToListAsync();
-
-            var customerDtos = customers.Select(c => CustomerMapper.EntityToDto(c)).ToList();
-
-            return new CustomerResponseDto
-            {
-                Items = customerDtos,
-                TotalCount = totalCount,
-                PageSize = pageSize,
-                CurrentPage = pageNumber
-            };
             
         }
 
-        public async Task<CustomerDto> GetCustomer(int id)
+        public async Task<CustomerDto> GetCustomer(int? id)
         {
-            var customer = await _context.Customers.FindAsync(id);
-            if (customer == null)
+            try
             {
-                return null;
+                if (id == null)
+                {
+                    return null;
+                }
+                else
+                {
+                    var customer = await _context.Customers.FindAsync(id);
+                    if (customer == null)
+                    {
+                        return null;
+                    }
+                    return CustomerMapper.EntityToDto(customer);
+                }
             }
-            return CustomerMapper.EntityToDto(customer);
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+
+                return null;
+
+            }
         }
         public async Task<CustomerDto> AddCustomer(CustomerDto customerDto)
         {
-            var customer = CustomerMapper.DtoToEntity(customerDto);
-
-            _context.Add(customer);
-            await _context.SaveChangesAsync();
-
-            return CustomerMapper.EntityToDto(customer);
-        }
-        public async Task<CustomerDto> UpdateCustomer(int id, CustomerUpdateDto customerDto)
-        {
-
-            var customer = CustomerMapper.DtoToEntity(customerDto);
-
-            _context.Entry(customer).State = EntityState.Modified;
-
             try
             {
+                var customer = CustomerMapper.DtoToEntity(customerDto);
+
+                _context.Add(customer);
                 await _context.SaveChangesAsync();
+
+                return CustomerMapper.EntityToDto(customer);
             }
-            catch (DbUpdateConcurrencyException)
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                return null;
+            }
+        }
+        public async Task<CustomerDto> UpdateCustomer(int? id, CustomerUpdateDto customerDto)
+        {
+
+            try
             {
                 if (!CustomerExists(id))
                 {
@@ -90,29 +115,63 @@ namespace mvp_onboarding.Server.Classes
                 }
                 else
                 {
-                    throw;
+                    var customer = CustomerMapper.DtoToEntity(customerDto);
+
+                    _context.Entry(customer).State = EntityState.Modified;
+                    await _context.SaveChangesAsync();
+                    return CustomerMapper.EntityToDto(customer);
                 }
             }
-
-            return CustomerMapper.EntityToDto(customer);
-
-        }
-        public async Task<CustomerDto> DeleteCustomer(int id)
-        {
-            var customer = await _context.Customers.FindAsync(id);
-            if (customer == null)
+            catch (Exception ex)
             {
+                Console.WriteLine($"Error: {ex.Message}");
+
                 return null;
             }
 
-            _context.Customers.Remove(customer);
-            await _context.SaveChangesAsync();
 
-            return CustomerMapper.EntityToDto(customer);
+
         }
-        public bool CustomerExists(int id)
+
+        public async Task<CustomerDto> DeleteCustomer(int? id)
         {
-            return _context.Customers.Any(e => e.Id == id);
+            try
+            {
+                if (id == null)
+                {
+                    return null;
+                }
+                else
+                {
+                    var customer = await _context.Customers.FindAsync(id);
+                    if (customer == null)
+                    {
+                        return null;
+                    }
+
+                    _context.Customers.Remove(customer);
+                    await _context.SaveChangesAsync();
+
+                    return CustomerMapper.EntityToDto(customer);
+                }
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                return null;
+            }
         }
+        public bool CustomerExists(int? id)
+        {
+            if (id == null)
+            {
+                return false;
+            }
+            else
+            {
+                return _context.Customers.Any(e => e.Id == id);
+            }
+        }
+
     }
 }
