+++ mvp_onboarding.Server/Classes/ProductMethods.cs
@@ -1,8 +1,8 @@
-﻿using mvp_onboarding.Server.Dtos;
+﻿using Microsoft.EntityFrameworkCore;
+using mvp_onboarding.Server.Dtos;
 using mvp_onboarding.Server.Interfaces;
-using mvp_onboarding.Server.Models;
-using Microsoft.EntityFrameworkCore;
 using mvp_onboarding.Server.Mappers;
+using mvp_onboarding.Server.Models;
 
 namespace mvp_onboarding.Server.Classes
 {
@@ -15,12 +15,11 @@ namespace mvp_onboarding.Server.Classes
 
         private readonly TalentOnboardingContext _context;
 
-        public async Task<ProductResponseDto> GetProducts(int pageNumber, int pageSize, string sortColumn, string sortDirection)
+        public async Task<ProductResponseDto> GetProducts(int pageNumber,int pageSize,string sortColumn,string sortDirection)
         {
-            var query = _context.Products.AsQueryable();
-
             try
             {
+                var query = _context.Products.AsQueryable();
                 if (sortDirection.ToLower() == "asc")
                 {
                     query = query.OrderBy(c => EF.Property<Product>(c, sortColumn));
@@ -29,59 +28,85 @@ namespace mvp_onboarding.Server.Classes
                 {
                     query = query.OrderByDescending(c => EF.Property<Product>(c, sortColumn));
                 }
+
+                var totalCount = await query.CountAsync();
+                var products = await query
+                    .Skip((pageNumber - 1) * pageSize)
+                    .Take(pageSize)
+                    .ToListAsync();
+
+                var productDtos = products.Select(c => ProductMapper.EntityToDto(c)).ToList();
+
+                return new ProductResponseDto
+                {
+                    Items = productDtos,
+                    TotalCount = totalCount,
+                    PageSize = pageSize,
+                    CurrentPage = pageNumber,
+                };
             }
             catch (InvalidOperationException ex)
             {
                 Console.WriteLine(ex.Message);
+
+                return new ProductResponseDto
+                {
+                    Items = new List<ProductDto>(),
+                    TotalCount = 0,
+                    PageSize = pageSize,
+                    CurrentPage = pageNumber,
+                };
             }
 
-            var totalCount = await query.CountAsync();
-            var products = await query
-                .Skip((pageNumber - 1) * pageSize)
-                .Take(pageSize)
-                .ToListAsync();
-
-            var productDtos = products.Select(c => ProductMapper.EntityToDto(c)).ToList();
-
-            return new ProductResponseDto
-            {
-                Items = productDtos,
-                TotalCount = totalCount,
-                PageSize = pageSize,
-                CurrentPage = pageNumber
-            };
         }
 
-        public async Task<ProductDto> GetProduct(int id)
+        public async Task<ProductDto> GetProduct(int? id)
         {
-            var product = await _context.Products.FindAsync(id);
-            if (product == null)
-            {
-                return null;
-            }
-            return ProductMapper.EntityToDto(product);
-        }
-        public async Task<ProductDto> AddProduct(ProductDto productDto)
-        {
-            var product = ProductMapper.DtoToEntity(productDto);
-
-            _context.Add(product);
-            await _context.SaveChangesAsync();
-
-            return ProductMapper.EntityToDto(product);
-        }
-        public async Task<ProductDto> UpdateProduct(int id, ProductUpdateDto productDto)
-        {
-
-            var product = ProductMapper.DtoToEntity(productDto);
-
-            _context.Entry(product).State = EntityState.Modified;
-
             try
             {
-                await _context.SaveChangesAsync();
+                if (id == null)
+                {
+                    return null;
+                }
+                else
+                {
+                    var product = await _context.Products.FindAsync(id);
+                    if (product == null)
+                    {
+                        return null;
+                    }
+                    return ProductMapper.EntityToDto(product);
+                }
             }
-            catch (DbUpdateConcurrencyException)
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+
+                return null;
+            }
+        }
+
+        public async Task<ProductDto> AddProduct(ProductDto productDto)
+        {
+            try
+            {
+                var product = ProductMapper.DtoToEntity(productDto);
+
+                _context.Add(product);
+                await _context.SaveChangesAsync();
+
+                return ProductMapper.EntityToDto(product);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                return null;
+            }
+        }
+
+        public async Task<ProductDto> UpdateProduct(int? id, ProductUpdateDto productDto)
+        {
+            try
             {
                 if (!ProductExists(id))
                 {
@@ -89,29 +114,61 @@ namespace mvp_onboarding.Server.Classes
                 }
                 else
                 {
-                    throw;
+                    var product = ProductMapper.DtoToEntity(productDto);
+                    _context.Entry(product).State = EntityState.Modified;
+
+                    await _context.SaveChangesAsync();
+                    return ProductMapper.EntityToDto(product);
                 }
             }
-
-            return ProductMapper.EntityToDto(product);
-
-        }
-        public async Task<ProductDto> DeleteProduct(int id)
-        {
-            var product = await _context.Products.FindAsync(id);
-            if (product == null)
+            catch (Exception ex)
             {
+                Console.WriteLine($"Error {ex.Message}");
+
                 return null;
             }
 
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync();
-
-            return ProductMapper.EntityToDto(product);
         }
-        public bool ProductExists(int id)
+
+        public async Task<ProductDto> DeleteProduct(int? id)
         {
-            return _context.Products.Any(e => e.Id == id);
+            try
+            {
+                if (id == null)
+                {
+                    return null;
+                }
+                else
+                {
+                    var product = await _context.Products.FindAsync(id);
+                    if (product == null)
+                    {
+                        return null;
+                    }
+
+                    _context.Products.Remove(product);
+                    await _context.SaveChangesAsync();
+
+                    return ProductMapper.EntityToDto(product);
+                }
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e.Message);
+                return null;
+            }
+        }
+
+        public bool ProductExists(int? id)
+        {
+            if (id == null)
+            {
+                return false;
+            }
+            else
+            {
+                return _context.Products.Any(e => e.Id == id);
+            }
         }
     }
 }
